requirements.txt
alembic==1.12.1
anndata==0.12.6
anthropic>=0.39.0
annotated-types==0.7.0
anyio==3.7.1
array-api-compat==1.12.0
bcrypt==4.3.0
certifi==2025.10.5
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
contourpy==1.3.3
crc32c==2.8
cryptography==45.0.5
cycler==0.12.1
donfig==0.8.1.post1
ecdsa==0.19.1
fastapi==0.104.1
fonttools==4.60.1
h11==0.16.0
h5py==3.15.1
httptools==0.6.4
idna==3.10
joblib==1.5.2
kiwisolver==1.4.9
legacy-api-wrap==1.5
llvmlite==0.45.1
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.7
natsort==8.4.0
networkx==3.5
numba==0.62.1
numcodecs==0.16.3
numpy==2.3.4
packaging==25.0
pandas==2.3.3
passlib==1.7.4
patsy==1.0.2
pillow==12.0.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.5.0
pydantic_core==2.14.1
pynndescent==0.5.13
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
rsa==4.9.1
scanpy==1.11.5
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
session-info2==0.2.3
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
statsmodels==0.14.5
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
umap-learn==0.5.9.post2
urllib3==2.5.0
uvicorn==0.24.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
wheel==0.45.1
zarr==3.1.3
